+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/SwiftUI/Calling/CallingViewComponent/ParticipantsListViewModel.swift
@@ -11,22 +11,29 @@ class ParticipantsListViewModel: ObservableObject {
     @Published var participantsList: [ParticipantsListCellViewModel] = []
     @Published var localParticipantsListCellViewModel: ParticipantsListCellViewModel
     var lastUpdateTimeStamp = Date()
+    private let localizationProvider: LocalizationProvider
 
-    init(localUserState: LocalUserState) {
-        localParticipantsListCellViewModel = ParticipantsListCellViewModel(localUserState: localUserState)
+    init(localUserState: LocalUserState,
+         localizationProvider: LocalizationProvider) {
+        localParticipantsListCellViewModel = ParticipantsListCellViewModel(localUserState: localUserState,
+                                                                           localizationProvider: localizationProvider)
+        self.localizationProvider = localizationProvider
     }
 
     func update(localUserState: LocalUserState,
                 remoteParticipantsState: RemoteParticipantsState) {
 
         if localParticipantsListCellViewModel.isMuted != (localUserState.audioState.operation == .off) {
-            localParticipantsListCellViewModel = ParticipantsListCellViewModel(localUserState: localUserState)
+            localParticipantsListCellViewModel = ParticipantsListCellViewModel(
+                localUserState: localUserState,
+                localizationProvider: localizationProvider)
         }
 
         if lastUpdateTimeStamp != remoteParticipantsState.lastUpdateTimeStamp {
             lastUpdateTimeStamp = remoteParticipantsState.lastUpdateTimeStamp
             participantsList = remoteParticipantsState.participantInfoList.map {
-                ParticipantsListCellViewModel(participantInfoModel: $0)
+                ParticipantsListCellViewModel(participantInfoModel: $0,
+                                              localizationProvider: localizationProvider)
             }
         }
     }