+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/SwiftUI/Calling/CallingViewModel.swift
@@ -11,6 +11,7 @@ class CallingViewModel: ObservableObject {
     @Published var isConfirmLeaveOverlayDisplayed: Bool = false
     @Published var isParticipantGridDisplayed: Bool
     let isRightToLeft: Bool
+    @Published var appState: AppStatus = .foreground
 
     private let compositeViewModelFactory: CompositeViewModelFactory
     private let logger: Logger
@@ -43,6 +44,7 @@ class CallingViewModel: ObservableObject {
         let isCallConnected = store.state.callingState.status == .connected
         let hasRemoteParticipants = store.state.remoteParticipantsState.participantInfoList.count > 0
         isParticipantGridDisplayed = isCallConnected && hasRemoteParticipants
+        appState = store.state.lifeCycleState.currentStatus
         controlBarViewModel = compositeViewModelFactory
             .makeControlBarViewModel(dispatchAction: actionDispatch, endCallConfirm: { [weak self] in
                 guard let self = self else {
@@ -58,11 +60,15 @@ class CallingViewModel: ObservableObject {
             }.store(in: &cancellables)
     }
 
+    func getLobbyOverlayViewModel() -> LobbyOverlayViewModel {
+        return compositeViewModelFactory.makeLobbyOverlayViewModel()
+    }
+
     // MARK: ConfirmLeaveOverlay
     func getLeaveCallButtonViewModel() -> PrimaryButtonViewModel {
         let leaveCallButtonViewModel = compositeViewModelFactory.makePrimaryButtonViewModel(
             buttonStyle: .primaryFilled,
-            buttonLabel: "Leave call",
+            buttonLabel: localizationProvider.getLocalizedString(.leaveCall),
             iconName: nil,
             isDisabled: false,
             action: { [weak self] in
@@ -78,7 +84,7 @@ class CallingViewModel: ObservableObject {
     func getCancelButtonViewModel() -> PrimaryButtonViewModel {
         let cancelButtonViewModel = compositeViewModelFactory.makePrimaryButtonViewModel(
             buttonStyle: .primaryOutline,
-            buttonLabel: "Cancel",
+            buttonLabel: localizationProvider.getLocalizedString(.cancel),
             iconName: nil,
             isDisabled: false,
             action: { [weak self] in
@@ -105,6 +111,10 @@ class CallingViewModel: ObservableObject {
     }
 
     func receive(_ state: AppState) {
+        if appState != state.lifeCycleState.currentStatus {
+            appState = state.lifeCycleState.currentStatus
+        }
+
         guard state.lifeCycleState.currentStatus == .foreground else {
             return
         }
@@ -114,7 +124,8 @@ class CallingViewModel: ObservableObject {
         infoHeaderViewModel.update(localUserState: state.localUserState,
                                    remoteParticipantsState: state.remoteParticipantsState)
         localVideoViewModel.update(localUserState: state.localUserState)
-        participantGridsViewModel.update(remoteParticipantsState: state.remoteParticipantsState)
+        participantGridsViewModel.update(remoteParticipantsState: state.remoteParticipantsState,
+                                         lifeCycleState: state.lifeCycleState)
         bannerViewModel.update(callingState: state.callingState)
         let isCallConnected = state.callingState.status == .connected
         let hasRemoteParticipants = state.remoteParticipantsState.participantInfoList.count > 0