+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/SwiftUI/Setup/SetupViewComponent/SetupControlBarViewModel.swift
@@ -11,6 +11,7 @@ class SetupControlBarViewModel: ObservableObject {
     @Published var audioPermission: AppPermission.Status = .unknown
     @Published var isAudioDeviceSelectionDisplayed: Bool = false
     private let logger: Logger
+    private let localizationProvider: LocalizationProvider
     private(set) var cameraButtonViewModel: IconWithLabelButtonViewModel!
     private(set) var micButtonViewModel: IconWithLabelButtonViewModel!
     private(set) var audioDeviceButtonViewModel: IconWithLabelButtonViewModel!
@@ -19,25 +20,27 @@ class SetupControlBarViewModel: ObservableObject {
     private var callingStatus: CallingStatus = .none
     private var cameraStatus: LocalUserState.CameraOperationalStatus = .off
     private(set) var micStatus: LocalUserState.AudioOperationalStatus = .off
-
     private var localVideoStreamId: String?
 
     private let dispatch: ActionDispatch
+    private var isJoinRequested: Bool = false
 
     init(compositeViewModelFactory: CompositeViewModelFactory,
          logger: Logger,
          dispatchAction: @escaping ActionDispatch,
-         localUserState: LocalUserState) {
+         localUserState: LocalUserState,
+         localizationProvider: LocalizationProvider) {
         self.logger = logger
         self.dispatch = dispatchAction
-
+        self.localizationProvider = localizationProvider
         self.audioDevicesListViewModel = compositeViewModelFactory.makeAudioDevicesListViewModel(
             dispatchAction: dispatchAction,
             localUserState: localUserState)
         self.cameraButtonViewModel = compositeViewModelFactory.makeIconWithLabelButtonViewModel(
             iconName: .videoOff,
             buttonTypeColor: .colorThemedWhite,
-            buttonLabel: "Video is off",
+            buttonLabel: self.localizationProvider
+                .getLocalizedString(.videoOff),
             isDisabled: false) { [weak self] in
                 guard let self = self else {
                     return
@@ -48,7 +51,8 @@ class SetupControlBarViewModel: ObservableObject {
         self.micButtonViewModel = compositeViewModelFactory.makeIconWithLabelButtonViewModel(
             iconName: .micOff,
             buttonTypeColor: .colorThemedWhite,
-            buttonLabel: "Mic is off",
+            buttonLabel: self.localizationProvider
+                .getLocalizedString(.micOff),
             isDisabled: false) { [weak self] in
                 guard let self = self else {
                     return
@@ -59,7 +63,8 @@ class SetupControlBarViewModel: ObservableObject {
         self.audioDeviceButtonViewModel = compositeViewModelFactory.makeIconWithLabelButtonViewModel(
             iconName: .speakerFilled,
             buttonTypeColor: .colorThemedWhite,
-            buttonLabel: "Device",
+            buttonLabel: self.localizationProvider
+                .getLocalizedString(.device),
             isDisabled: false) { [weak self] in
                 guard let self = self else {
                     return
@@ -102,11 +107,15 @@ class SetupControlBarViewModel: ObservableObject {
     }
 
     func isCameraDisabled() -> Bool {
-        cameraPermission == .denied
+        return isJoinRequested || cameraPermission == .denied
     }
 
     func isAudioDisabled() -> Bool {
-        audioPermission == .denied
+        return isJoinRequested || audioPermission == .denied
+    }
+
+    func isControlBarHidden() -> Bool {
+        return audioPermission == .denied
     }
 
     func update(localUserState: LocalUserState,
@@ -131,14 +140,26 @@ class SetupControlBarViewModel: ObservableObject {
         audioDevicesListViewModel.update(audioDeviceStatus: localUserState.audioState.device)
     }
 
+    func update(isJoinRequested: Bool) {
+        self.isJoinRequested = isJoinRequested
+        cameraButtonViewModel.update(isDisabled: isCameraDisabled())
+        micButtonViewModel.update(isDisabled: isAudioDisabled())
+        audioDeviceButtonViewModel.update(isDisabled: isJoinRequested)
+    }
+
     private func updateButtonViewModel(localUserState: LocalUserState) {
         cameraButtonViewModel.update(
             iconName: self.cameraStatus == .on ? .videoOn : .videoOff,
-            buttonLabel: "Video is \(self.cameraStatus == .on ? "on" : "off")")
+            buttonLabel: self.cameraStatus == .on
+            ? self.localizationProvider.getLocalizedString(.videoOn)
+            : self.localizationProvider.getLocalizedString(.videoOff))
         cameraButtonViewModel.update(isDisabled: isCameraDisabled())
         micButtonViewModel.update(
             iconName: self.micStatus == .on ? .micOn : .micOff,
-            buttonLabel: "Mic is \(self.micStatus == .on ? "on" : "off")")
+            buttonLabel: self.micStatus == .on
+            ? self.localizationProvider.getLocalizedString(.micOn)
+            : self.localizationProvider.getLocalizedString(.micOff))
+        micButtonViewModel.update(isDisabled: isAudioDisabled())
         audioDeviceButtonViewModel.update(
             iconName: deviceIconFor(audioDeviceStatus: localUserState.audioState.device),
             buttonLabel: deviceLabelFor(audioDeviceStatus: localUserState.audioState.device))
@@ -170,13 +191,13 @@ class SetupControlBarViewModel: ObservableObject {
     private func deviceLabelFor(audioDeviceStatus: LocalUserState.AudioDeviceSelectionStatus) -> String {
         switch audioDeviceStatus {
         case .bluetoothSelected:
-            return AudioDeviceType.bluetooth.rawValue
+            return localizationProvider.getLocalizedString(AudioDeviceType.bluetooth.name)
         case .headphonesSelected:
-            return AudioDeviceType.headphones.name
+            return localizationProvider.getLocalizedString(AudioDeviceType.headphones.name)
         case .receiverSelected:
-            return AudioDeviceType.receiver.name
+            return localizationProvider.getLocalizedString(AudioDeviceType.receiver.name)
         case .speakerSelected:
-            return AudioDeviceType.speaker.name
+            return localizationProvider.getLocalizedString(AudioDeviceType.speaker.name)
         default:
             return audioDeviceButtonViewModel.buttonLabel
         }