+++ AzureCommunicationUI/AzureCommunicationUITests/CompositeViewModelFactoryMocking.swift
@@ -21,6 +21,7 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     var previewAreaViewModel: PreviewAreaViewModel?
     var setupControlBarViewModel: SetupControlBarViewModel?
     var errorInfoViewModel: ErrorInfoViewModel?
+    var lobbyOverlayViewModel: LobbyOverlayViewModel?
     var audioDevicesListViewModel: AudioDevicesListViewModel?
     var primaryButtonViewModel: PrimaryButtonViewModel?
     var iconButtonViewModel: IconButtonViewModel?
@@ -93,14 +94,18 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     func makeAudioDevicesListViewModel(dispatchAction: @escaping ActionDispatch,
                                        localUserState: LocalUserState) -> AudioDevicesListViewModel {
         return audioDevicesListViewModel ?? AudioDevicesListViewModel(dispatchAction: dispatchAction,
-                                                                      localUserState: localUserState)
+                                                                      localUserState: localUserState,
+                                                                      localizationProvider: LocalizationProviderMocking())
     }
 
     func makeErrorInfoViewModel() -> ErrorInfoViewModel {
-        return errorInfoViewModel ?? ErrorInfoViewModel()
+        return errorInfoViewModel ?? ErrorInfoViewModel(localizationProvider: LocalizationProviderMocking())
     }
 
     // MARK: CallingViewModels
+    func makeLobbyOverlayViewModel() -> LobbyOverlayViewModel {
+        return lobbyOverlayViewModel ?? LobbyOverlayViewModel(localizationProvider: LocalizationProviderMocking())
+    }
     func makeControlBarViewModel(dispatchAction: @escaping ActionDispatch,
                                  endCallConfirm: @escaping (() -> Void),
                                  localUserState: LocalUserState) -> ControlBarViewModel {
@@ -114,7 +119,8 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     func makeInfoHeaderViewModel(localUserState: LocalUserState) -> InfoHeaderViewModel {
         return infoHeaderViewModel ?? InfoHeaderViewModel(compositeViewModelFactory: self,
                                                           logger: logger,
-                                                          localUserState: localUserState)
+                                                          localUserState: localUserState,
+                                                          localizationProvider: LocalizationProviderMocking())
     }
 
     func makeParticipantCellViewModel(participantModel: ParticipantInfoModel) -> ParticipantGridCellViewModel {
@@ -127,7 +133,8 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     }
 
     func makeParticipantsListViewModel(localUserState: LocalUserState) -> ParticipantsListViewModel {
-        return participantsListViewModel ?? ParticipantsListViewModel(localUserState: localUserState)
+        return participantsListViewModel ?? ParticipantsListViewModel(localUserState: localUserState,
+                                                                      localizationProvider: LocalizationProviderMocking())
     }
 
     func makeBannerViewModel() -> BannerViewModel {
@@ -135,12 +142,14 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     }
 
     func makeBannerTextViewModel() -> BannerTextViewModel {
-        return bannerTextViewModel ?? BannerTextViewModel()
+        return bannerTextViewModel ?? BannerTextViewModel(localizationProvider: LocalizationProviderMocking())
     }
 
     // MARK: SetupViewModels
     func makePreviewAreaViewModel(dispatchAction: @escaping ActionDispatch) -> PreviewAreaViewModel {
-        return previewAreaViewModel ?? PreviewAreaViewModel(compositeViewModelFactory: self, dispatchAction: dispatchAction)
+        return previewAreaViewModel ?? PreviewAreaViewModel(compositeViewModelFactory: self,
+                                                            dispatchAction: dispatchAction,
+                                                            localizationProvider: LocalizationProviderMocking())
     }
 
     func makeSetupControlBarViewModel(dispatchAction: @escaping ActionDispatch,
@@ -148,6 +157,11 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
         return setupControlBarViewModel ?? SetupControlBarViewModel(compositeViewModelFactory: self,
                                                                     logger: logger,
                                                                     dispatchAction: dispatchAction,
-                                                                    localUserState: localUserState)
+                                                                    localUserState: localUserState,
+                                                                    localizationProvider: LocalizationProviderMocking())
+    }
+
+    func makeJoiningCallActivityViewModel() -> JoiningCallActivityViewModel {
+        JoiningCallActivityViewModel(localizationProvider: LocalizationProviderMocking())
     }
 }