+++ AzureCommunicationUI/AzureCommunicationUITests/Mocking/ViewModels/ParticipantGridViewModelMocking.swift
@@ -7,15 +7,15 @@ import Foundation
 @testable import AzureCommunicationUI
 
 class ParticipantGridViewModelMocking: ParticipantGridViewModel {
-    private let updateState: ((RemoteParticipantsState) -> Void)?
+    private let updateState: ((RemoteParticipantsState, LifeCycleState) -> Void)?
 
     init(compositeViewModelFactory: CompositeViewModelFactory,
-         updateState: ((RemoteParticipantsState) -> Void)? = nil) {
+         updateState: ((RemoteParticipantsState, LifeCycleState) -> Void)? = nil) {
         self.updateState = updateState
         super.init(compositeViewModelFactory: compositeViewModelFactory)
     }
 
-    override func update(remoteParticipantsState: RemoteParticipantsState) {
-        updateState?(remoteParticipantsState)
+    override func update(remoteParticipantsState: RemoteParticipantsState, lifeCycleState: LifeCycleState) {
+        updateState?(remoteParticipantsState, lifeCycleState)
     }
 }